Add a basic implementation of the A2A bridge for local use
This version of the bridge lacks works well locally when integrated
with an IDE or coding agent, but is not yet suitable for deployment
into a cluster because it lacks the ability to handle identity
delegation.

To do this, the impersonnate user capability should be leveraged. Where
callers may send a request with a token proving their kubernetes
identity, and those permissions will be inhereted by the MCP server.